PR issue solved

+++ src/controller/cartController.js
@@ -10,12 +10,7 @@ exports.addToCart = async (req, res) => {
 
     // If User is avalable then Added Proudcuts to same cart other wise create new cart
     if (isUserAvailable) {
-      const update = {
-        $push: {
-          products: products,
-        },
-      };
-      await cartModel.updateOne({ userId: userId }, update);
+      await cartModel.updateOne({ userId }, { $push: { products } });
     } else {
       await cartModel.create({ userId, products });
     }
@@ -30,7 +25,7 @@ exports.addToCart = async (req, res) => {
 exports.removeFromCart = async (req, res) => {
   try {
     // Delete data from cart using cart id
-    await cartModel.findOneAndDelete(req.params.userId);
+    await cartModel.findOneAndDelete({ userId: req.params.userId });
     res.status(200).send("data deleted sucessfully from cart");
   } catch (err) {
     res.status(400).send(err.message + "Data Deletion Unsucessufl from cart");
@@ -43,9 +38,8 @@ exports.findCart = async (req, res) => {
     const cartData = await cartModel
       .findOne({ userId: req.params.userId })
       .lean();
-    res.status(201).send(cartData.products);
+    res.status(200).send(cartData);
   } catch (err) {
     res.send(err.message + "Fetching data ").status(400);
   }
 };
-const user = require("../models/user");
